requirements.txt
convex
resend
astral
requests
python-dotenv
beautifulsoup4
